refactor(annotation): move appContextDecorator into src/testHelpers/
Relocate the Storybook decorator out of .storybook/ into
src/testHelpers/ so it lives with the package's other test
infrastructure. Drop the unconditional global decorator in preview.tsx
in favor of meta-level opt-in — each story file declares which
context wrapper it wants. Prepares for extracting the Storybook
config into its own workspace, where there is no shared "this is the
annotation Storybook" assumption to lean on.
